// src/app/models/candidate-info.module.ts
export class CandidateInfoModel {
  bssId: number;
  candidateName: string;
  phoneNo: string;
  position: string;
  email: string;
  primarySkill: string;
  secondarySkill: string;
  resume: string;
  dob: Date;
  currentLocation: string;
  currentCompany: string;
  noticePeriod: string;
  totalExp: string;
  releventExp: string;
  currentCtc: string;
  expectedCtc: string;
  reasonForChange: string;
  readyToRelocate: number;
  createdAt: Date;
  updatedAt: Date;
}